+++ src/main/java/com/axperty/farmingexperiencecore/event/AntiFarmEvents.java
@@ -1,6 +1,7 @@
 package com.axperty.farmingexperiencecore.event;
 
 import com.axperty.farmingexperiencecore.FarmingExperienceCore;
+import com.axperty.farmingexperiencecore.config.ModConfig;
 import net.minecraft.world.entity.Mob;
 import net.minecraft.world.entity.monster.EnderMan;
 import net.minecraft.world.entity.player.Player;
@@ -16,7 +17,7 @@ public class AntiFarmEvents {
 
     @SubscribeEvent
     public static void onLivingDrops(LivingDropsEvent event) {
-        if (event.getEntity() instanceof Mob) {
+        if (ModConfig.enableActiveLootingOnly && event.getEntity() instanceof Mob) {
             if (!(event.getSource().getEntity() instanceof Player)) {
                 event.setCanceled(true);
             }
@@ -26,10 +27,10 @@ public class AntiFarmEvents {
     @SubscribeEvent
     public static void onExperienceDrop(LivingExperienceDropEvent event) {
         if (event.getEntity() instanceof Mob mob) {
-            if (event.getAttackingPlayer() == null) {
+            if (ModConfig.enableActiveLootingOnly && event.getAttackingPlayer() == null) {
                 event.setDroppedExperience(0);
                 event.setCanceled(true);
-            } else if (mob instanceof EnderMan) {
+            } else if (ModConfig.enableEndermanXpNerf && mob instanceof EnderMan) {
                 // Nerf Enderman XP to 2 to discourage XP farms
                 event.setDroppedExperience(2);
             }