+++ changelog.md
@@ -7,17 +7,20 @@
 Some features in this new version have been inspired from the [Matcha Flavoured](https://modrinth.com/datapack/matcha-flavoured) datapack by [klei_wright](https://modrinth.com/user/klei_wright).
 
 - The player will not lose their inventory items on death, instead, they will lose hearts that can be recovered by eating golden apples.
-- Iron golems and phantoms will not spawn anymore.
+- Phantoms will not spawn anymore.
 - Zombies are scared of any light sources and will run away from them.
 - Zombies, skeletons, and endermans spawn in the overworld.
 - Creepers and spiders spawn in caves.
 - All monsters have less health.
+- Mobs only drop loot and XP when directly killed by a player to discourage passive mob farms.
+- Endermen always drop a pearl and drop less XP to discourage enderman farms.
 - Untamed cats don't run away from the player.
 - TNT can be thrown when the player has flint and steel and TNT blocks on their hands.
 - Minecarts can be placed without using rails.
 - Multijump, dash, and smash enchantments.
 - Tools and weapons can be always repaired.
 - Days and nights are longer.
+- Configuration file and configuration screen to edit whatever the player wants with support for Cloth Config API.
 
 ## 🛠️ Minor Changes: